Type cast to integers

+++ 0x00-hello_world/6-size.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 /**
- * main - xxxxxxx
+ * main - print the size of various types
  *
  * Description: prints the size of various types on the computer it is compiled
  * and run on
@@ -11,11 +11,11 @@
 
 int main(void)
 {
-	printf("Size of a char: %u byte(s)\n", sizeof(char));
-	printf("Size of an int: %u byte(s)\n", sizeof(int));
-	printf("Size of a long int: %u byte(s)\n", sizeof(long int));
-	printf("Size of a long long int: %u byte(s)\n", sizeof(long long int));
-	printf("Size of a float: %u byte(s)\n", sizeof(float));
+	printf("Size of a char: %d byte(s)\n", (int) sizeof(char));
+	printf("Size of an int: %d byte(s)\n", (int) sizeof(int));
+	printf("Size of a long int: %d byte(s)\n", (int) sizeof(long int));
+	printf("Size of a long long int: %d byte(s)\n", (int) sizeof(long long int));
+	printf("Size of a float: %d byte(s)\n", (int) sizeof(float));
 
 	return (0);
 }
